requirements.txt
fastapi
uvicorn
jinja2
python-multipart

chromadb
langchain
langchain-groq

pdfplumber
docx2txt

python-dotenv
